Change to using BigDecimals

+++ src/main/java/sweep/ISaving.java
@@ -1,5 +1,7 @@
 package sweep;
 
+import java.math.BigDecimal;
+
 /**
  * ISaving.
  *
@@ -12,6 +14,6 @@ public interface ISaving {
      * 
      * @return
      */
-    int getAmmount();
+    BigDecimal getAmmount();
 
 }
